# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hamqadam AI Identity Verification",
    "description": "Face, document and fraud analysis for identity verification. Returns a recommendation; the Backend owns the decision.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/verify": {
      "post": {
        "tags": [
          "verification"
        ],
        "summary": "Run a full identity verification",
        "description": "Analyse one submission and return a recommendation.\n\nEvery image is optional at the transport layer, deliberately. A missing\none is a finding the pipeline reports rather than a request the API\nrefuses - the Backend usually still wants the analysis of whatever did\narrive.\n\nNote:\n    The multipart body is parsed here rather than declared as `Form`/\n    `File` parameters, and the schema Swagger renders comes from\n    ``openapi_extra`` above. Two problems forced that, and neither is\n    solvable with plain parameter annotations:\n\n    1. Swagger's \"Send empty value\" checkbox, ticked by default on\n       optional file fields, posts an empty **string** for a field the\n       user left blank. Declared as ``UploadFile | None`` FastAPI\n       refuses the whole request with \"Expected UploadFile, received:\n       <class 'str'>\" - so the documentation page could not submit its\n       own form.\n    2. Widening the annotation to ``UploadFile | str`` fixed that and\n       broke something worse: Swagger renders a union as a *text box*,\n       so `secondary_images` became ``array<(string | string)>`` with no\n       file picker at all. The upload UI was traded for the ability to\n       submit.\n\n    Parsing the form directly gives both: the schema below declares\n    honest ``format: binary`` fields, so Swagger shows real file\n    pickers and a repeatable selector for `secondary_images`, while the\n    parser treats an empty part exactly as an absent one.",
        "operationId": "verify_v1_verify_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "verification_id": {
                    "type": "string",
                    "description": "Your identifier for this attempt. Echoed back and used as the log correlation id.",
                    "example": "ver_01HQ8XZ"
                  },
                  "user_reference": {
                    "type": "string",
                    "description": "Your account identifier. Pseudonymised before it reaches a log and never echoed in the response. Also the key the duplicate gallery stores against."
                  },
                  "enrol_on_success": {
                    "type": "boolean",
                    "description": "Optional override. Leave unset and the configured duplicate.enrol_policy decides - this endpoint verifies, checks the gallery and enrols in one call. Pass false to suppress enrolment for this submission. Needs user_reference; without one there is no key to store against. A REJECT is never enrolled."
                  },
                  "live_selfie": {
                    "type": "string",
                    "format": "binary",
                    "description": "The biometric reference for the whole decision."
                  },
                  "profile_image": {
                    "type": "string",
                    "format": "binary",
                    "description": "The account's main photograph."
                  },
                  "cnic_image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Front of the CNIC."
                  },
                  "secondary_images": {
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "type": "array",
                    "description": "Additional photographs of the same person."
                  }
                },
                "type": "object",
                "required": [
                  "verification_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/duplicate/{reference}": {
      "delete": {
        "tags": [
          "gallery"
        ],
        "summary": "Erase a face from the duplicate gallery",
        "description": "Erase a stored template.\n\nIdempotent: erasing an absent reference succeeds. A caller retrying an\nerasure request must not be told it failed the second time, because\nthat is how erasure requests get abandoned half-done.",
        "operationId": "forget_v1_duplicate__reference__delete",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reference"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Forget V1 Duplicate  Reference  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/qdrant/count": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "How many templates the gallery holds (dev only)",
        "description": "Gallery size and which adapter answered.\n\n`store` is the field worth reading: if it says `memory` the gallery is\nnot durable, whatever the configuration claims.",
        "operationId": "admin_count_admin_qdrant_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Admin Count Admin Qdrant Count Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/admin/qdrant/list": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Enumerate stored references (dev only)",
        "description": "List references, without vectors.\n\nBounded by `admin.max_list_limit`: a gallery can hold hundreds of\nthousands of records, and an unbounded scan is both a memory problem and\na bulk-disclosure one.",
        "operationId": "admin_list_admin_qdrant_list_get",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Admin List Admin Qdrant List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "Per-caller API key issued by the Backend. Compared in constant time; the key itself is never written to a log.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}